Moved books model within the module
