+++ src/tests/books.test.js
@@ -7,7 +7,7 @@ import Faker from 'faker';
 import server from '../index';
 import User from '../app/modules/auth/user.model';
 import Author from '../app/modules/authors/authors.model';
-import Book from '../app/models/book';
+import Book from '../app/modules/books/books.model';
 import { createToken } from '../app/config/authentication/jwt';
 import confirmationLevels from '../app/constants/confirmation_levels';
